create dedicated file test
